requirements.txt
fastapi
uvicorn
pydantic
motor
twilio
python-dotenv
dnspython